[PAYMENT-METHOD]revamping payment method components
