+++ src/reactapp/src/components/login/context/LoginFormContext.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const LoginFormContext = React.createContext();
+
+export default LoginFormContext;
