+++ src/reactapp/src/components/login/hooks/useLoginFormContext.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import LoginFormContext from '../context/LoginFormContext';
+
+export default function useLoginFormContext() {
+  return useContext(LoginFormContext);
+}
