+++ src/reactapp/src/components/paymentMethod/components/NoPaymentMethodInfoBox.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function NoPaymentMethodInfoBox() {
+  return (
+    <div className="h-32 py-4 min-h-12">
+      <div className="flex items-center justify-center w-full h-full">
+        <div>No payment methods available at the moment</div>
+      </div>
+    </div>
+  );
+}
+
+export default NoPaymentMethodInfoBox;
