+++ src/reactapp/src/components/paymentMethod/components/PaymentMethodFormManager.jsx
@@ -1,15 +1,14 @@
-import React, { useCallback, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { node } from 'prop-types';
 import _get from 'lodash.get';
 import { string as YupString } from 'yup';
 import { Form, useFormikContext } from 'formik';
 
-import PaymentMethodFormContext from './PaymentMethodFormContext';
+import PaymentMethodFormContext from '../context/PaymentMethodFormContext';
 import useFormSection from '../../../hook/useFormSection';
+import usePaymentMethodAppContext from '../hooks/usePaymentMethodAppContext';
+import usePaymentMethodCartContext from '../hooks/usePaymentMethodCartContext';
 import { PAYMENT_METHOD_FORM } from '../../../config';
-import useAppContext from '../../../hook/useAppContext';
-import setPaymentMethod from '../../../api/cart/setPaymentMethod';
-import useCartContext from '../../../hook/useCartContext';
 
 const initialValues = {
   code: '',
@@ -22,26 +21,34 @@ const validationSchema = {
 };
 
 function PaymentMethodFormManager({ children }) {
-  const [, { setPageLoader }] = useAppContext();
-  const { selectedPaymentMethod } = useCartContext();
   const { setFieldValue } = useFormikContext();
+  const {
+    selectedPaymentMethod,
+    setPaymentMethod,
+  } = usePaymentMethodCartContext();
+  const {
+    setPageLoader,
+    setSuccessMessage,
+    setErrorMessage,
+  } = usePaymentMethodAppContext();
 
-  const formSubmit = useCallback(
-    async values => {
-      const paymentMethodSelected = _get(values, PAYMENT_METHOD_FORM);
+  const formSubmit = async values => {
+    const paymentMethodSelected = _get(values, PAYMENT_METHOD_FORM);
 
-      try {
-        if (paymentMethodSelected.code) {
-          setPageLoader(true);
-          await setPaymentMethod(paymentMethodSelected);
-          setPageLoader(false);
-        }
-      } catch (error) {
+    try {
+      if (paymentMethodSelected.code) {
+        setPageLoader(true);
+        await setPaymentMethod(paymentMethodSelected);
+        setSuccessMessage('Payment method added successfully.');
         setPageLoader(false);
       }
-    },
-    [setPageLoader]
-  );
+    } catch (error) {
+      setPageLoader(false);
+      setErrorMessage(
+        'Something went wrong while adding payment to the quote.'
+      );
+    }
+  };
 
   useEffect(() => {
     if (selectedPaymentMethod.code) {
