+++ src/reactapp/src/components/paymentMethod/hooks/usePaymentMethodFormContext.js
@@ -1,6 +1,5 @@
 import { useContext } from 'react';
-
-import PaymentMethodFormContext from '../context/Form/PaymentMethod/PaymentMethodFormContext';
+import PaymentMethodFormContext from '../context/PaymentMethodFormContext';
 
 export default function usePaymentMethodFormContext() {
   return useContext(PaymentMethodFormContext);
