+++ src/reactapp/src/components/paymentMethod/index.js
@@ -1,0 +1,3 @@
+import PaymentMethod from './PaymentMethod';
+
+export default PaymentMethod;
