--- src/reactapp/src/context/Form/GuestEmail/GuestEmailFormContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const GuestEmailFormContext = React.createContext();
-
-export default GuestEmailFormContext;
