+++ src/reactapp/src/context/Form/index.js
@@ -1,11 +1,7 @@
-import GuestEmailMan from './GuestEmail/GuestEmailFormManager';
 import FormStepProvider from './Step/StepProvider';
 
-import GuestEmailFormCxt from './GuestEmail/GuestEmailFormContext';
 import FormStepCxt from './Step/StepContext';
 
-export const GuestEmailFormManager = GuestEmailMan;
 export const StepProvider = FormStepProvider;
 
-export const GuestEmailFormContext = GuestEmailFormCxt;
 export const StepContext = FormStepCxt;
